requirements.txt
# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM & AI - Compatible versions
langchain>=1.1.0
langchain-core>=1.1.0
langchain-community>=0.3.15
langchain-groq>=1.1.0
langchain-huggingface>=1.1.0
langsmith>=0.4.0

# Sentence Transformers & NLP
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# HuggingFace Integration
huggingface-hub>=0.19.0
accelerate>=0.24.0
safetensors>=0.4.0

# Database (MongoDB)
pymongo>=4.6.0

# Data Handling
pandas>=2.0.0
numpy>=1.24.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# NLP Utilities
textblob>=0.17.0

# Additional utilities
python-dateutil>=2.8.0


# # Core Framework
# streamlit>=1.28.0
# python-dotenv>=1.0.0

# # LLM & AI
# langchain>=0.3.0
# langchain-groq>=0.1.0
# langchain-huggingface>=0.0.1
# sentence-transformers>=2.2.2
# transformers>=4.35.0
# torch>=2.0.0

# # HuggingFace Integration
# huggingface-hub>=0.19.0
# accelerate>=0.24.0

# # Database (MongoDB)
# pymongo>=4.6.0

# # Data Handling
# pandas>=2.0.0

# # Data Validation
# pydantic>=2.0.0

# # Visualization
# plotly>=5.17.0
# matplotlib>=3.7.0
# seaborn>=0.12.0

# # NLP
# textblob>=0.17.0
